+++ mods/Mining/mod/data.lua
@@ -1,2 +1,4 @@
 require("prototypes.item.items")
+require("prototypes.item.cables")
+
 require("prototypes.group")